extrator_telefone_empresas/requirements.txt
pandas
openpyxl
requests
beautifulsoup4
python-dotenv
tqdm
pytest
